requirements.txt
flask==3.0.3
line-bot-sdk==3.14.2
requests==2.32.3
python-dotenv==1.0.1
firebase-admin==6.6.0
google-cloud==0.34.0
google-cloud-storage==2.19.0
google-auth==2.26.2

